# openapi.yaml
openapi: 3.0.0
info:
  title: API Fizjoterapeutycznej Aplikacji
  version: 1.0.0
  description: Dokumentacja podstawowych endpointów z poprawionymi schematami na podstawie Mongoose models
servers:
  - url: http://localhost:8080
    description: Serwer lokalny
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Patient:
      type: object
      properties:
        _id:
          type: string
        personalInfo:
          type: object
          properties:
            firstName:
              type: string
            lastName:
              type: string
            pesel:
              type: string
            dateOfBirth:
              type: string
              format: date
            gender:
              type: string
              enum: [M, F, Other]
            contact:
              type: object
              properties:
                phone:
                  type: string
                email:
                  type: string
                  format: email
        medicalInfo:
          type: object
          properties:
            allergies:
              type: array
              items: { type: string }
            chronicConditions:
              type: array
              items: { type: string }
            medications:
              type: array
              items:
                type: object
                properties:
                  name: { type: string }
                  dosage: { type: string }
                  frequency: { type: string }
            medicalHistory:
              type: array
              items: { type: string }
        physiotherapyData:
          type: object
          properties:
            referringPhysician: { type: string }
            initialDiagnosis: { type: string }
            treatmentGoals:
              type: array
              items: { type: string }
            contraindications:
              type: array
              items: { type: string }
            specialNotes: { type: string }
        consentGiven:
          type: boolean
        consentDate:
          type: string
          format: date-time
    Employee:
      type: object
      properties:
        _id:
          type: string
        personalInfo:
          type: object
          properties:
            firstName: { type: string }
            lastName: { type: string }
            contact:
              type: object
              properties:
                phone: { type: string }
                email: { type: string, format: email }
        roles:
          type: array
          items: { type: string }
        permissions:
          type: array
          items: { type: string }
        status:
          type: string
          enum: [active, inactive]
        schedule:
          type: array
          items:
            type: object
            properties:
              date: { type: string, format: date }
              slots:
                type: array
                items:
                  type: object
                  properties:
                    startTime: { type: string }
                    endTime: { type: string }
    Appointment:
      type: object
      properties:
        _id: { type: string }
        patient:
          $ref: '#/components/schemas/Patient'
        employee:
          $ref: '#/components/schemas/Employee'
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [scheduled, confirmed, completed, cancelled]
        notes:
          type: string
    Examination:
      type: object
      properties:
        _id: { type: string }
        patient:
          $ref: '#/components/schemas/Patient'
        employee:
          $ref: '#/components/schemas/Employee'
        subjectiveAssessment:
          type: object
          properties:
            chiefComplaint: { type: string }
            painScale: { type: integer, minimum: 0, maximum: 10 }
            historyOfPresentIllness: { type: string }
            pastMedicalHistory: { type: string }
            socialHistory: { type: string }
        objectiveAssessment:
          type: object
          properties:
            observation: { type: string }
            palpation: { type: string }
            rangeOfMotion: { type: string }
            muscleStrength: { type: string }
            functionalTests: { type: string }
            specialTests: { type: string }
            neurological: { type: string }
        assessment:
          type: object
          properties:
            primaryDiagnosis: { type: string }
            secondaryDiagnoses:
              type: array
              items: { type: string }
            prognosisAssessment: { type: string }
        plan:
          type: object
          properties:
            treatmentPlan: { type: string }
            goals:
              type: object
              properties:
                shortTerm:
                  type: array
                  items: { type: string }
                longTerm:
                  type: array
                  items: { type: string }
            contraindications:
              type: array
              items: { type: string }
            recommendations: { type: string }
        status:
          type: string
          enum: [draft, completed, reviewed]
    Visit:
      type: object
      properties:
        _id: { type: string }
        patient:
          $ref: '#/components/schemas/Patient'
        employee:
          $ref: '#/components/schemas/Employee'
        appointment:
          $ref: '#/components/schemas/Appointment'
        date:
          type: string
          format: date-time
        status:
          type: string
          enum: [scheduled, in-progress, completed, cancelled]
        notes: { type: string }
        treatment:
          type: object
          additionalProperties: true
...
security:
  - bearerAuth: []
paths:
...
  /api/examinations/{id}:
    get:
      tags: [Examinations]
      summary: Pobieranie pojedynczego badania
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Dane badania
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Examination'
    put:
      tags: [Examinations]
      summary: Aktualizacja badania
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Examination'
      responses:
        '200':
          description: Badanie zaktualizowane
    patch:
      tags: [Examinations]
      summary: Zakończenie badania
      operationId: completeExamination
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Badanie zakończone
    patch:
      tags: [Examinations]
      summary: Weryfikacja badania przez specjalistę
      operationId: reviewExamination
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Badanie zweryfikowane
...
  /api/visits/{id}:
    get:
      tags: [Visits]
      summary: Pobieranie pojedynczej wizyty
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Dane wizyty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
    put:
      tags: [Visits]
      summary: Aktualizacja wizyty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Visit'
      responses:
        '200':
          description: Wizyta zaktualizowana
    patch:
      tags: [Visits]
      summary: Zakończenie wizyty
      operationId: completeVisit
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Wizyta zakończona
    patch:
      tags: [Visits]
      summary: Anulowanie wizyty
      operationId: cancelVisit
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Wizyta anulowana
...